Routine in notes, fixed bugs with routine creation
