+++ lib/chat_page.dart
@@ -1,23 +1,23 @@
 import 'dart:async';
 import 'dart:developer';
 
-import 'package:Chrono/models/tag.dart';
-import 'package:Chrono/record.service.dart';
-import 'package:dart_openai/dart_openai.dart';
-import 'package:flutter/material.dart';
 import 'package:Chrono/colors.dart';
+import 'package:Chrono/features/notes/data/db/record.service.dart';
+import 'package:Chrono/helpers/api-key-options.dart';
 import 'package:Chrono/helpers/token.helper.dart';
 import 'package:Chrono/message_bubble.dart';
 import 'package:Chrono/message_composer.dart';
 import 'package:Chrono/models/record.dart';
+import 'package:Chrono/models/tag.dart';
 import 'package:Chrono/services/gpt-note-bind.service.dart';
 import 'package:Chrono/services/gpt.service.dart';
 import 'package:Chrono/services/messages.service.dart';
 import 'package:Chrono/shared/instructions-block.dart';
 import 'package:Chrono/shared/instructions.dart';
+import 'package:dart_openai/dart_openai.dart';
+import 'package:flutter/material.dart';
 
-import 'package:Chrono/helpers/api-key-options.dart';
-import 'db_manager.dart';
+import 'core/db/db_manager.dart';
 import 'models/chat-message.dart';
 
 class ChatPage extends StatefulWidget {
@@ -33,7 +33,7 @@ class _ChatPageState extends State<ChatPage> {
   final _messages = <ChatMessage>[];
   final List<Record> allRecords = [];
   List<Tag> allTags = [];
-  List<ChatMessage> _tokenMessages = [];
+  // List<ChatMessage> _tokenMessages = [];
 
   var _awaitingResponse = false;
   var includeAllNote = false;
