+++ lib/core/ui/widgets/bottom_nav_bar/notes_bottom_nav_bar.dart
@@ -6,6 +6,7 @@ import 'package:go_router/go_router.dart';
 import '../../../../colors.dart';
 import '../../../../features/start/presentation/cubit/start_cubit.dart';
 import '../../../../features/start/presentation/cubit/start_state.dart';
+import '../../../../generated/l10n.dart';
 
 class ChronoBottomNavigationBar extends StatelessWidget {
   const ChronoBottomNavigationBar({super.key, required this.navigationShell});
@@ -24,24 +25,25 @@ class ChronoBottomNavigationBar extends StatelessWidget {
         onTap: (index) {
           context.read<StartCubit>().changeBottomNavBar(index);
           navigationShell.goBranch(index);
+          print('INDEX == $index');
         },
         items: [
           BottomNavigationBarItem(
             icon: const Icon(
               Icons.note,
             ),
-            label: 'Notes',
+            label: S.of(context).notes,
           ),
           BottomNavigationBarItem(
             icon: const Icon(Icons.cell_tower),
-            label: 'Health Center',
+            label: S.of(context).productivity_center,
           ),
           BottomNavigationBarItem(
             icon: SvgPicture.asset(
               'assets/icons/chat.svg',
               colorFilter: ColorFilter.mode(MyColors.forthyColor, BlendMode.srcIn),
             ),
-            label: 'GPT',
+            label: S.of(context).gpt,
           ),
         ],
       );
