+++ lib/features/health_reminder/presentation/bloc/health_reminder_state.dart
@@ -5,16 +5,16 @@ class HealthReminderState extends Equatable {
   final String id;
   final bool needExit;
   final bool isLoading;
-  final List<RepeatMode> modes;
   final List<DayOfWeek> daysOfWeek;
+  final List<String> selectedDays;
 
   HealthReminderState({
     required this.date,
     this.id = '',
     this.needExit = false,
     this.isLoading = false,
-    required this.modes,
     required this.daysOfWeek,
+    required this.selectedDays,
   });
 
   HealthReminderState copyWith({
@@ -22,16 +22,16 @@ class HealthReminderState extends Equatable {
     String? id,
     bool? needExit,
     bool? isLoading,
-    List<RepeatMode>? modes,
     List<DayOfWeek>? daysOfWeek,
+    List<String>? selectedDays,
   }) {
     return HealthReminderState(
       date: date ?? this.date,
       id: id ?? this.id,
       needExit: needExit ?? this.needExit,
       isLoading: isLoading ?? this.isLoading,
-      modes: modes ?? this.modes,
       daysOfWeek: daysOfWeek ?? this.daysOfWeek,
+      selectedDays: selectedDays ?? this.selectedDays,
     );
   }
 
@@ -41,7 +41,7 @@ class HealthReminderState extends Equatable {
         id,
         needExit,
         isLoading,
-        modes,
         daysOfWeek,
+        selectedDays,
       ];
 }
