+++ lib/features/notes/di/notes_module.dart
@@ -1,0 +1,6 @@
+import 'package:Chrono/core/di/dependency_injection.dart';
+import 'package:Chrono/features/notes/data/db/record.service.dart';
+
+void initNotesModule() {
+  getIt.registerFactory(() => RecordService());
+}
