+++ lib/features/notes/presentation/bloc/home_event.dart
@@ -1,0 +1,11 @@
+part of 'home_bloc.dart';
+
+sealed class HomeEvent extends Equatable {
+  @override
+  List<Object?> get props => [];
+}
+
+class HomeStreamDataChanged extends HomeEvent {
+  @override
+  List<Object?> get props => [];
+}
