+++ lib/instuction_page.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/material.dart';
-import 'package:Chrono/db_manager.dart';
+import 'package:Chrono/core/db/db_manager.dart';
 import 'package:Chrono/models/instructions.model.dart';
+import 'package:flutter/material.dart';
 
 class InstructionsPage extends StatefulWidget {
   @override
