+++ lib/l10n/intl_ru.arb
@@ -1,4 +1,7 @@
 {
+  "notes": "Notes",
+  "productivity_center": "Productivity Center",
+  "gpt": "GPT",
   "repeat": "Repeat",
   "enter_text": "Enter a description",
   "description": "Description",
