+++ lib/shared/instructions-block.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/material.dart';
-import 'package:Chrono/db_manager.dart';
+import 'package:Chrono/core/db/db_manager.dart';
 import 'package:Chrono/models/instructions.model.dart';
+import 'package:flutter/material.dart';
 
 class InstructionsBlockWidget extends StatefulWidget {
   final Function(String) onSubmitted;
